Add ability to specify custom selection highlight group in setup function
Previously, the selection highlight group was hardcoded to "Visual". This commit allows users to specify a custom highlight group using the `hlgroup` option in the `setup` function. The new variable `hlgroup` is initialized in the `setup` function and used in the `gpt_replace_opfunc` function to highlight the selected text. This provides more flexibility and customization for users who want to use a different highlight group for selection.